+++ changelog.md
@@ -6,6 +6,7 @@ Core:
 
 - Allow to leave workspace
 - Allow to kick employee
+- Allow to create employee
 
 HR: